Cloud Funtions/songs_classification/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.41.0
pandas==1.3.0
requests==2.26.0
sklearn==0.0